feat: clickhouse handler no longer support insert with format.
