+++ src/query/ast/src/ast/format/ast_format.rs
@@ -1141,30 +1141,6 @@ impl<'ast> Visitor<'ast> for AstFormatVisitor {
 
     fn visit_insert_source(&mut self, insert_source: &'ast InsertSource) {
         match insert_source {
-            InsertSource::Streaming { format, .. } => {
-                let streaming_name = format!("StreamSource {}", format);
-                let streaming_format_ctx = AstFormatContext::new(streaming_name);
-                let streaming_node = FormatTreeNode::new(streaming_format_ctx);
-                self.children.push(streaming_node);
-            }
-            InsertSource::StreamingV2 { settings, .. } => {
-                let mut file_formats_children = Vec::new();
-                for (k, v) in settings.options.iter() {
-                    let file_format_name = format!("FileFormat {} = {:?}", k, v);
-                    let file_format_format_ctx = AstFormatContext::new(file_format_name);
-                    let file_format_node = FormatTreeNode::new(file_format_format_ctx);
-                    file_formats_children.push(file_format_node);
-                }
-                let file_formats_format_name = "StreamSourceFileFormats".to_string();
-                let files_formats_format_ctx = AstFormatContext::with_children(
-                    file_formats_format_name,
-                    file_formats_children.len(),
-                );
-                let files_formats_node =
-                    FormatTreeNode::with_children(files_formats_format_ctx, file_formats_children);
-
-                self.children.push(files_formats_node);
-            }
             InsertSource::Values { .. } => {
                 let values_name = "ValueSource".to_string();
                 let values_format_ctx = AstFormatContext::new(values_name);
