--- src/query/pipeline/sources/src/input_formats/mod.rs
@@ -1,35 +1,0 @@
-// Copyright 2021 Datafuse Labs
-//
-// Licensed under the Apache License, Version 2.0 (the "License");
-// you may not use this file except in compliance with the License.
-// You may obtain a copy of the License at
-//
-//     http://www.apache.org/licenses/LICENSE-2.0
-//
-// Unless required by applicable law or agreed to in writing, software
-// distributed under the License is distributed on an "AS IS" BASIS,
-// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-// See the License for the specific language governing permissions and
-// limitations under the License.
-
-mod beyond_end_reader;
-pub mod error_utils;
-mod impls;
-mod input_context;
-mod input_format;
-mod input_format_text;
-mod input_pipeline;
-mod input_split;
-mod source_aligner;
-mod transform_deserializer;
-
-pub use beyond_end_reader::BeyondEndReader;
-pub use input_context::InputContext;
-pub use input_context::InputPlan;
-pub use input_context::StreamPlan;
-pub use input_format::InputFormat;
-pub use input_format_text::*;
-pub use input_pipeline::StreamingReadBatch;
-pub use input_split::split_by_size;
-pub use input_split::SplitInfo;
-pub use source_aligner::Aligner;
