+++ src/query/storages/stage/src/lib.rs
@@ -19,9 +19,8 @@
 #![allow(clippy::uninlined_format_args)]
 
 mod append;
-mod input_context_bridge;
-mod one_file_partition;
 mod read;
 mod stage_table;
+mod utils;
 
 pub use stage_table::StageTable;
