+++ src/query/storages/stage/src/read/row_based/formats/ndjson/block_builder.rs
@@ -20,7 +20,6 @@ use databend_common_expression::ColumnBuilder;
 use databend_common_expression::DataBlock;
 use databend_common_formats::FieldJsonAstDecoder;
 use databend_common_meta_app::principal::NullAs;
-use databend_common_pipeline_sources::input_formats::error_utils::truncate_column_data;
 use databend_common_storage::FileParseError;
 
 use crate::read::load_context::LoadContext;
@@ -28,6 +27,7 @@ use crate::read::row_based::batch::RowBatchWithPosition;
 use crate::read::row_based::format::RowDecoder;
 use crate::read::row_based::formats::ndjson::format::NdJsonInputFormat;
 use crate::read::row_based::processors::BlockBuilderState;
+use crate::read::row_based::utils::truncate_column_data;
 
 pub struct NdJsonDecoder {
     pub load_context: Arc<LoadContext>,
