+++ src/query/storages/stage/src/read/row_based/formats/tsv/block_builder.rs
@@ -20,7 +20,6 @@ use databend_common_expression::ColumnBuilder;
 use databend_common_expression::DataBlock;
 use databend_common_formats::SeparatedTextDecoder;
 use databend_common_io::cursor_ext::BufferReadStringExt;
-use databend_common_pipeline_sources::input_formats::error_utils::get_decode_error_by_pos;
 use databend_common_storage::FileParseError;
 
 use crate::read::load_context::LoadContext;
@@ -28,6 +27,7 @@ use crate::read::row_based::batch::RowBatchWithPosition;
 use crate::read::row_based::format::RowDecoder;
 use crate::read::row_based::formats::tsv::format::TsvInputFormat;
 use crate::read::row_based::processors::BlockBuilderState;
+use crate::read::row_based::utils::get_decode_error_by_pos;
 
 pub struct TsvDecoder {
     pub load_context: Arc<LoadContext>,
